% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\encoding{UTF-8}
\name{glwd_get}
\alias{glwd_get}
\title{glwd_get}
\usage{
glwd_get(level)
}
\arguments{
\item{level}{integer data level}
}
\description{
Retrieve each level of the Global Lakes and Wetlands
Database.
}
\examples{
\dontrun{
glwd_get(level = 1)
}
}
\references{
Lehner, B., Döll, P.: Development and validation of a global
database of lakes, reservoirs and wetlands, Journal of Hydrology, Volume
296, Issues 1–4, 20 August 2004, Pages 1-22,
http://dx.doi.org/10.1016/j.jhydrol.2004.03.028.
}

