% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{glwd_load}
\alias{glwd_load}
\title{glwd_load}
\usage{
glwd_load(level)
}
\arguments{
\item{level}{integer data level}
}
\description{
Load each level of the Global Lakes and Wetlands
 Database into an R object.
}
\examples{
\dontrun{
dt <- glwd_load(level = 1)
}
}

